# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Returns Services API",
        "description": "API endpoints of devgp project.",
        "version": "v1"
    },
    "servers": [
        {
            "url": "https://devgp.test/api/v1",
            "description": "Api v1"
        }
    ],
    "paths": {
        "/groups": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "get list of the groups",
                "operationId": "2fd561b1c25f8a824ddb5ee57d0591e4",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "query",
                        "description": "slug of a tag",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "order by which column(members-views-daily_views) e.g. <b>views,desc</b>",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "array of groups returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GroupModel"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Groups"
                ],
                "summary": "create new group",
                "operationId": "58b0a9cd406b84d7ea7c9c6f78844a39",
                "parameters": [
                    {
                        "name": "image",
                        "in": "path",
                        "description": "image file of the group",
                        "required": false,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreGroupRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "group details returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GroupModel"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "form validation error"
                    }
                }
            }
        },
        "/groups/{slug}": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "get a group by slug",
                "operationId": "cab47bed8b8e60d809836241c292e42a",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "The slug of group",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "group details returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GroupModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "group not found"
                    }
                }
            }
        },
        "/groups/{slug}/related": {
            "get": {
                "tags": [
                    "Groups"
                ],
                "summary": "get list of the related groups",
                "operationId": "922199e164ab64856475e222db68baf3",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "The slug of group",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "The count of related groups",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 10,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "array of related groups returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GroupModel"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tags/{slug}": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "get a tag by slug",
                "operationId": "20cd15038f54ea54358cbd4ceb966d96",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "The slug of tag",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "group details returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TagModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "tag not found"
                    }
                }
            }
        },
        "/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "search between tags (use in create-group form)",
                "operationId": "ad9608b8a52558850a1927426d8e712b",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "maximum number of results",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "array of tags returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TagModel"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "GroupModel": {
                "title": "Group",
                "properties": {
                    "name": {
                        "description": ".",
                        "type": "string",
                        "example": "گروه برنامه نویسان لاراول در تلگرام"
                    },
                    "slug": {
                        "description": "english name of the group for url",
                        "type": "string",
                        "example": "laravel"
                    },
                    "image": {
                        "description": ".",
                        "type": "string",
                        "example": "https://site.com/full/path/to/laravel.jpg"
                    },
                    "description": {
                        "description": ".",
                        "type": "string",
                        "example": ""
                    },
                    "address": {
                        "description": ".",
                        "type": "string",
                        "format": "uri",
                        "example": "https://t.me/laravel"
                    },
                    "members": {
                        "description": "number of group members",
                        "type": "number",
                        "format": "int64",
                        "example": "2000"
                    },
                    "views": {
                        "description": "page views in our site",
                        "type": "number",
                        "format": "int64",
                        "example": "100"
                    }
                },
                "type": "object"
            },
            "StoreGroupRequest": {
                "required": [
                    "name",
                    "address"
                ],
                "properties": {
                    "name": {
                        "description": ".",
                        "type": "string",
                        "example": "گروه برنامه نویسان لاراول در تلگرام"
                    },
                    "slug": {
                        "description": "english name of the group for url",
                        "type": "string",
                        "example": "laravel"
                    },
                    "description": {
                        "description": ".",
                        "type": "string",
                        "example": "لورم ایپسوم متن ساختگی..."
                    },
                    "address": {
                        "description": ".",
                        "type": "string",
                        "format": "uri",
                        "example": "https://t.me/laravel"
                    },
                    "tags": {
                        "description": "array of tag slugs",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": "['laravel', 'backend']"
                    }
                },
                "type": "object"
            },
            "TagModel": {
                "title": "Tag",
                "properties": {
                    "name": {
                        "description": ".",
                        "type": "string",
                        "example": "جاوا اسکریپت or javascript"
                    },
                    "title": {
                        "description": "\\<title\\> tag for seo",
                        "type": "string",
                        "example": "گروه های برنامه نویسی جاوا اسکریپت"
                    },
                    "slug": {
                        "description": "english name of the tag for url",
                        "type": "string",
                        "example": "javascript"
                    },
                    "description": {
                        "description": "for top of tag page, to improve seo. most of times is null btw :))",
                        "type": "string",
                        "example": ""
                    },
                    "color": {
                        "description": ".",
                        "type": "string",
                        "example": "#ff0000"
                    }
                },
                "type": "object"
            }
        }
    }
}